Fixed incorrect value type in ct_item_t constructor
Now you can actually create lists with non-integer item values, using the
normal item() method, or ct_item_t constructor.

+++ src/gfxengine/toolkit.cpp
@@ -251,7 +251,7 @@ void ct_label_t::render()
 	ct_item_t::
 ----------------------------------------------------------*/
 
-ct_item_t::ct_item_t(const char *cap, int val)
+ct_item_t::ct_item_t(const char *cap, double val)
 {
 	if(cap)
 		caption(cap);
